pyproject.toml
[tool.poetry]
    name        = "bell-avr-vmc-thermal-module"
    version     = "0.1.0"
    description = "Thermal camera to MQTT module for Bell AVR"
    license     = "MIT"
    readme      = "README.md"
    homepage    = "https://github.com/bellflight/AVR-VMC-Thermal-Module"
    authors     = ["Nathan Vaughn <nvaughn51@gmail.com>"]

[tool.poetry.dependencies]
    python                         = ">=3.9,<3.12"
    adafruit-circuitpython-amg88xx = "1.2.18"
    loguru                         = "^0.6.0"
    bell-avr-libraries             = "0.2.0a0"
    # https://github.com/NVIDIA/jetson-gpio/issues/79
    jetson-gpio                    = { version = "2.0.21", markers = "sys_platform == 'linux'" }

[tool.poetry.group.dev.dependencies]
    pre-commit  = "^2.21.0"
    pytest      = "^7.2.1"
    pytest-mock = "^3.10.0"
    pytest-cov  = "^4.0.0"

[tool.pyright]
    typeCheckingMode = "basic"
    venvPath         = "."
    venv             = ".venv"

[tool.ruff]
    ignore = ["E501"]
